// src/app/interfaces/seguridad/empresa.interface.ts
export interface EmpresaRequest {
    ruc: string;
    razonSocial: string;
    nombreComercial: string;
    tipoEmpresa: string;
    representante: string;
    dniRepresentante: string;
    direccion: string;
    telefono: string;
    email: string;
    paginaWeb: string;
    logoUrl: string;
    codigoDepartamento: string;
    codigoProvincia: string;
    codigoDistrito: string;
}